backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# Database
sqlalchemy
asyncpg
alembic
redis

# ML & Image Processing
tensorflow
opencv-python-headless
pillow
numpy
scikit-learn

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

# AWS Integration
boto3

# Messaging (Twilio for SMS/WhatsApp)
twilio
httpx

# Utilities
aiofiles
python-magic
